+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/tyanotifications/tya/NotificationsItBase.java
@@ -43,7 +43,6 @@ import uk.gov.hmcts.reform.sscs.thirdparty.pdfservice.DocmosisPdfService;
 import uk.gov.hmcts.reform.sscs.tyanotifications.config.NotificationConfig;
 import uk.gov.hmcts.reform.sscs.tyanotifications.config.NotificationTestRecipients;
 import uk.gov.hmcts.reform.sscs.tyanotifications.controller.NotificationController;
-import uk.gov.hmcts.reform.sscs.tyanotifications.controller.ReminderTestController;
 import uk.gov.hmcts.reform.sscs.tyanotifications.factory.NotificationFactory;
 import uk.gov.hmcts.reform.sscs.tyanotifications.service.*;
 import uk.gov.hmcts.reform.sscs.tyanotifications.service.docmosis.PdfLetterService;
